Refactor `theme_bloc`
