+++ lib/bootstrap.dart
@@ -11,7 +11,7 @@ import 'package:taskify/core/services/notification_service.dart';
 import 'package:taskify/features/auth/blocs/login_bloc/login_bloc.dart';
 import 'package:taskify/features/auth/blocs/register_bloc/register_bloc.dart';
 import 'package:taskify/features/auth/blocs/reset_pwd_bloc/reset_pwd_bloc.dart';
-import 'package:taskify/features/profile/blocs/theme_bloc.dart';
+import 'package:taskify/features/profile/blocs/theme_bloc/theme_bloc.dart';
 import 'package:taskify/firebase_options.dart';
 
 class AppBlocObserver extends BlocObserver {
